dev.requirements.txt
pytest-django==3.7.0
pre-commit==1.20.0
black==19.10b0
isort==4.3.21
flake8==3.7.9
pytest-testdox==1.2.1
